+++ api/typescript/src/core/kurtosis_core_rpc_api_bindings/api_container_service_pb.d.ts
@@ -601,6 +601,81 @@ export namespace KurtosisInstructionPosition {
   }
 }
 
+export class KurtosisError extends jspb.Message {
+  getInterpretationError(): KurtosisInterpretationError | undefined;
+  setInterpretationError(value?: KurtosisInterpretationError): KurtosisError;
+  hasInterpretationError(): boolean;
+  clearInterpretationError(): KurtosisError;
+
+  getValidationError(): KurtosisValidationError | undefined;
+  setValidationError(value?: KurtosisValidationError): KurtosisError;
+  hasValidationError(): boolean;
+  clearValidationError(): KurtosisError;
+
+  getExecutionError(): KurtosisExecutionError | undefined;
+  setExecutionError(value?: KurtosisExecutionError): KurtosisError;
+  hasExecutionError(): boolean;
+  clearExecutionError(): KurtosisError;
+
+  getErrorCase(): KurtosisError.ErrorCase;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): KurtosisError.AsObject;
+  static toObject(includeInstance: boolean, msg: KurtosisError): KurtosisError.AsObject;
+  static serializeBinaryToWriter(message: KurtosisError, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): KurtosisError;
+  static deserializeBinaryFromReader(message: KurtosisError, reader: jspb.BinaryReader): KurtosisError;
+}
+
+export namespace KurtosisError {
+  export type AsObject = {
+    interpretationError?: KurtosisInterpretationError.AsObject,
+    validationError?: KurtosisValidationError.AsObject,
+    executionError?: KurtosisExecutionError.AsObject,
+  }
+
+  export enum ErrorCase { 
+    ERROR_NOT_SET = 0,
+    INTERPRETATION_ERROR = 1,
+    VALIDATION_ERROR = 2,
+    EXECUTION_ERROR = 3,
+  }
+}
+
+export class KurtosisExecutionResponseLine extends jspb.Message {
+  getInstruction(): KurtosisInstruction | undefined;
+  setInstruction(value?: KurtosisInstruction): KurtosisExecutionResponseLine;
+  hasInstruction(): boolean;
+  clearInstruction(): KurtosisExecutionResponseLine;
+
+  getError(): KurtosisError | undefined;
+  setError(value?: KurtosisError): KurtosisExecutionResponseLine;
+  hasError(): boolean;
+  clearError(): KurtosisExecutionResponseLine;
+
+  getKurtosisExecutionResponseLineCase(): KurtosisExecutionResponseLine.KurtosisExecutionResponseLineCase;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): KurtosisExecutionResponseLine.AsObject;
+  static toObject(includeInstance: boolean, msg: KurtosisExecutionResponseLine): KurtosisExecutionResponseLine.AsObject;
+  static serializeBinaryToWriter(message: KurtosisExecutionResponseLine, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): KurtosisExecutionResponseLine;
+  static deserializeBinaryFromReader(message: KurtosisExecutionResponseLine, reader: jspb.BinaryReader): KurtosisExecutionResponseLine;
+}
+
+export namespace KurtosisExecutionResponseLine {
+  export type AsObject = {
+    instruction?: KurtosisInstruction.AsObject,
+    error?: KurtosisError.AsObject,
+  }
+
+  export enum KurtosisExecutionResponseLineCase { 
+    KURTOSIS_EXECUTION_RESPONSE_LINE_NOT_SET = 0,
+    INSTRUCTION = 1,
+    ERROR = 2,
+  }
+}
+
 export class StartServicesArgs extends jspb.Message {
   getServiceIdsToConfigsMap(): jspb.Map<string, ServiceConfig>;
   clearServiceIdsToConfigsMap(): StartServicesArgs;