+++ internal_testsuites/typescript/src/testsuite/startosis_upload_files_test/startsosis_upload_files.test.ts
@@ -28,8 +28,8 @@ PATH_TO_MOUNT_UPLOADED_DIR = "` + PATH_TO_MOUNT_UPLOADED_DIR + `"
 
 print("Adding service " + DATASTORE_SERVICE_ID + ".")
 
-uploaded_artifact_uuid = upload_files(DIR_TO_UPLOAD)
-print("Uploaded " + uploaded_artifact_uuid)
+uploaded_artifact_id = upload_files(DIR_TO_UPLOAD)
+print("Uploaded " + uploaded_artifact_id)
 
 
 config = struct(
@@ -38,7 +38,7 @@ config = struct(
         DATASTORE_PORT_ID: struct(number = DATASTORE_PORT_NUMBER, protocol = DATASTORE_PORT_PROTOCOL)
     },
 	files = {
-		uploaded_artifact_uuid: PATH_TO_MOUNT_UPLOADED_DIR
+		uploaded_artifact_id: PATH_TO_MOUNT_UPLOADED_DIR
 	}
 )